refactor: redesign project structure
