+++ src/index.js
@@ -5,7 +5,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 
 
-import App from "./components/app";
+import App from "./app";
 
 
 ReactDOM.render(
